타이틀 버그 해결

+++ NWT-Dungeon/NWT-Dungeon/SelectEnemyTurn.cpp
@@ -10,7 +10,7 @@ SelectEnemyTurn::SelectEnemyTurn(StateMachine* stateMachine) {
 }
 
 void SelectEnemyTurn::UpdateState() {
-	if (ChooseIndex(0, m_enemies->size() - 1, true, m_core->renderer->select)) {
+	if (ChooseIndex(0, (int)m_enemies->size() - 1, true, m_core->renderer->select)) {
 		m_core->GetSelectedPlayer()->SetTarget((*m_enemies)[m_core->renderer->select]);
 		m_core->renderer->select = 0;
